@@ -1,11 +1,18 @@
 <?php
 
-namespace ByJG\AnyDataset\Db\Helpers;
+namespace ByJG\AnyDataset\Db\SqlDialect;
 
-use ByJG\AnyDataset\Db\DbDriverInterface;
+use ByJG\AnyDataset\Core\Exception\DatabaseException;
+use ByJG\AnyDataset\Db\DatabaseExecutor;
+use ByJG\AnyDataset\Db\Exception\DbDriverNotConnected;
 use ByJG\AnyDataset\Db\IsolationLevelEnum;
+use ByJG\AnyDataset\Db\SqlStatement;
+use ByJG\XmlUtil\Exception\FileException;
+use ByJG\XmlUtil\Exception\XmlUtilException;
+use Override;
+use Psr\SimpleCache\InvalidArgumentException;
 
-class DbOci8Functions extends DbBaseFunctions
+class OciDialect extends BaseSqlDialect
 {
 
     public function __construct()
@@ -16,6 +23,7 @@ class DbOci8Functions extends DbBaseFunctions
         $this->deliTableRight = '"';
     }
 
+    #[Override]
     public function concat(string $str1, ?string $str2 = null): string
     {
         return implode(' || ', func_get_args());
@@ -28,17 +36,19 @@ class DbOci8Functions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
+    #[Override]
     public function limit(string $sql, int $start, int $qty = 50): string
     {
         if (stripos($sql, ' OFFSET ') === false && stripos($sql, ' FETCH NEXT ') === false) {
-            $sql = $sql . " OFFSET x ROWS FETCH NEXT y ROWS ONLY";
+            return $sql . " OFFSET $start ROWS FETCH NEXT $qty ROWS ONLY";
         }
 
-        return preg_replace(
+        $result = preg_replace(
             '~(\s[Oo][Ff][Ff][Ss][Ee][Tt])\s.*?\s([Rr][Oo][Ww][Ss])\s.*?\s([Ff][Ee][Tt][Cc][Hh]\s[Nn][Ee][Xx][Tt])\s.*~',
             '$1 ' . $start . ' $2 ' . '$3 ' . $qty . ' ROWS ONLY',
             $sql
         );
+        return $result !== null ? $result : $sql;
     }
 
     /**
@@ -47,6 +57,7 @@ class DbOci8Functions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
+    #[Override]
     public function top(string $sql, int $qty): string
     {
         return $this->limit($sql, 0, $qty);
@@ -56,6 +67,7 @@ class DbOci8Functions extends DbBaseFunctions
      * Return if the database provider have a top or similar function
      * @return bool
      */
+    #[Override]
     public function hasTop(): bool
     {
         return true;
@@ -65,6 +77,7 @@ class DbOci8Functions extends DbBaseFunctions
      * Return if the database provider have a limit function
      * @return bool
      */
+    #[Override]
     public function hasLimit(): bool
     {
         return true;
@@ -78,6 +91,7 @@ class DbOci8Functions extends DbBaseFunctions
      * @return string
      * @example $db->getDbFunctions()->SQLDate("d/m/Y H:i", "dtcriacao")
      */
+    #[Override]
     public function sqlDate(string $format, ?string $column = null): string
     {
         if (is_null($column)) {
@@ -109,46 +123,53 @@ class DbOci8Functions extends DbBaseFunctions
     }
 
     /**
-     * @param DbDriverInterface $dbdataset
-     * @param string $sql
+     * @param DatabaseExecutor $executor
+     * @param string|SqlStatement $sql
      * @param array|null $param
      * @return mixed
+     * @throws DatabaseException
+     * @throws DbDriverNotConnected
+     * @throws FileException
+     * @throws XmlUtilException
+     * @throws InvalidArgumentException
      */
-    public function executeAndGetInsertedId(DbDriverInterface $dbdataset, string $sql, ?array $param = null): mixed
+    #[Override]
+    public function executeAndGetInsertedId(DatabaseExecutor $executor, string|SqlStatement $sql, ?array $param = null): mixed
     {
-        preg_match('/INSERT INTO ([a-zA-Z0-9_]+)/i', $sql, $matches);
+        $sqlString = $sql instanceof SqlStatement ? $sql->getSql() : $sql;
+        preg_match('/INSERT INTO ([a-zA-Z0-9_]+)/i', $sqlString, $matches);
         $tableName = $matches[1] ?? null;
 
         if (!empty($tableName)) {
             $tableName = strtoupper($tableName);
 
             // Get the primary key of the table
-            $primaryKeyResult = $dbdataset->getScalar("SELECT cols.column_name
+            $primaryKeyResult = $executor->getScalar("SELECT cols.column_name
                 FROM all_constraints cons, all_cons_columns cols
-                WHERE cols.table_name = '{$tableName}'
+                WHERE cols.table_name = '$tableName'
                 AND cons.constraint_type = 'P'
                 AND cons.constraint_name = cols.constraint_name
                 AND cons.owner = cols.owner
                 AND ROWNUM = 1");
 
             // Get the default value of the primary key
-            $defaultValueResult = $dbdataset->getScalar("SELECT DATA_DEFAULT
+            $defaultValueResult = $executor->getScalar("SELECT DATA_DEFAULT
                 FROM USER_TAB_COLUMNS
-                WHERE TABLE_NAME = '{$tableName}'
-                AND COLUMN_NAME = '{$primaryKeyResult}'");
+                WHERE TABLE_NAME = '$tableName'
+                AND COLUMN_NAME = '$primaryKeyResult'");
         }
 
-        $dbdataset->execute($sql, $param);
+        $executor->execute($sql, $param);
 
         if (!empty($tableName) && !empty($defaultValueResult)) {
 
             // Check if the default value is a sequence's nextval
-            if (strpos($defaultValueResult, '.nextval') !== false) {
+            if (str_contains($defaultValueResult, '.nextval')) {
                 // Extract the sequence name
                 $sequenceName = str_replace('.nextval', '', $defaultValueResult);
 
                 // Return the CURRVAL of the sequence
-                return $dbdataset->getScalar("SELECT {$sequenceName}.currval FROM DUAL");
+                return $executor->getScalar("SELECT $sequenceName.currval FROM DUAL");
             }
         }
 
@@ -156,12 +177,14 @@ class DbOci8Functions extends DbBaseFunctions
 
     }
 
+    #[Override]
     public function hasForUpdate(): bool
     {
         return true;
     }
 
-    public function getTableMetadata(DbDriverInterface $dbdataset, string $tableName): array
+    #[Override]
+    public function getTableMetadata(DatabaseExecutor $executor, string $tableName): array
     {
         $tableName = strtoupper($tableName);
         $sql = "SELECT
@@ -175,12 +198,13 @@ class DbOci8Functions extends DbBaseFunctions
                         END AS TYPE,
                     DATA_DEFAULT AS COLUMN_DEFAULT,
                     NULLABLE
-                FROM ALL_TAB_COLUMNS WHERE TABLE_NAME = '{$tableName}'";
+                FROM ALL_TAB_COLUMNS WHERE TABLE_NAME = '$tableName'";
 
-        return $this->getTableMetadataFromSql($dbdataset, $sql);
+        return $this->getTableMetadataFromSql($executor, $sql);
     }
 
-    protected function parseColumnMetadata($metadata)
+    #[Override]
+    protected function parseColumnMetadata(array $metadata): array
     {
         $return = [];
 
@@ -189,13 +213,14 @@ class DbOci8Functions extends DbBaseFunctions
                     'name' => $value['column_name'],
                     'dbType' => strtolower($value['type']),
                     'required' => $value['nullable'] == 'N',
-                    'default' => isset($value['column_default']) ? $value['column_default'] : null,
+                    'default' => $value['column_default'] ?? null,
                 ] + $this->parseTypeMetadata(strtolower($value['type']));
         }
 
         return $return;
     }
 
+    #[Override]
     public function getIsolationLevelCommand(?IsolationLevelEnum $isolationLevel = null): string
     {
         return match ($isolationLevel) {
