@@ -1,12 +1,13 @@
 <?php
 
-namespace ByJG\AnyDataset\Db\Helpers;
+namespace ByJG\AnyDataset\Db\SqlDialect;
 
 use ByJG\AnyDataset\Core\Exception\NotAvailableException;
-use ByJG\AnyDataset\Db\DbDriverInterface;
+use ByJG\AnyDataset\Db\DatabaseExecutor;
 use ByJG\AnyDataset\Db\IsolationLevelEnum;
+use Override;
 
-class DbSqliteFunctions extends DbBaseFunctions
+class SqliteDialect extends BaseSqlDialect
 {
 
     public function __construct()
@@ -22,6 +23,7 @@ class DbSqliteFunctions extends DbBaseFunctions
      * @param string|null $str2
      * @return string
      */
+    #[Override]
     public function concat(string $str1, ?string $str2 = null): string
     {
         return implode(' || ', func_get_args());
@@ -34,17 +36,19 @@ class DbSqliteFunctions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
+    #[Override]
     public function limit(string $sql, int $start, int $qty = 50): string
     {
         if (stripos($sql, ' LIMIT ') === false) {
-            $sql = $sql . " LIMIT x, y";
+            return $sql . " LIMIT $start, $qty";
         }
 
-        return preg_replace(
+        $result = preg_replace(
             '~(\s[Ll][Ii][Mm][Ii][Tt])\s.*?,\s*.*~',
             '$1 ' . $start .', ' .$qty,
             $sql
         );
+        return $result !== null ? $result : $sql;
     }
 
     /**
@@ -53,6 +57,7 @@ class DbSqliteFunctions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
+    #[Override]
     public function top(string $sql, int $qty): string
     {
         return $this->limit($sql, 0, $qty);
@@ -62,6 +67,7 @@ class DbSqliteFunctions extends DbBaseFunctions
      * Return if the database provider have a top or similar function
      * @return bool
      */
+    #[Override]
     public function hasTop(): bool
     {
         return true;
@@ -71,6 +77,7 @@ class DbSqliteFunctions extends DbBaseFunctions
      * Return if the database provider have a limit function
      * @return bool
      */
+    #[Override]
     public function hasLimit(): bool
     {
         return true;
@@ -84,6 +91,7 @@ class DbSqliteFunctions extends DbBaseFunctions
      * @return string
      * @example $db->getDbFunctions()->SQLDate("d/m/Y H:i", "dtcriacao")
      */
+    #[Override]
     public function sqlDate(string $format, ?string $column = null): string
     {
         if (is_null($column)) {
@@ -116,6 +124,7 @@ class DbSqliteFunctions extends DbBaseFunctions
         );
     }
 
+    #[Override]
     public function getSqlLastInsertId(): string
     {
         return "select last_insert_rowid() id";
@@ -126,23 +135,27 @@ class DbSqliteFunctions extends DbBaseFunctions
      * @return string
      * @throws NotAvailableException
      */
+    #[Override]
     public function forUpdate(string $sql): string
     {
         throw new NotAvailableException('FOR UPDATE not available for SQLite');
     }
 
+    #[Override]
     public function hasForUpdate(): bool
     {
         return false;
     }
 
-    public function getTableMetadata(DbDriverInterface $dbdataset, string $tableName): array
+    #[Override]
+    public function getTableMetadata(DatabaseExecutor $executor, string $tableName): array
     {
         $sql = "PRAGMA table_info(" . $this->deliTableLeft . $tableName . $this->deliTableRight . ")";
-        return $this->getTableMetadataFromSql($dbdataset, $sql);
+        return $this->getTableMetadataFromSql($executor, $sql);
     }
 
-    protected function parseColumnMetadata($metadata)
+    #[Override]
+    protected function parseColumnMetadata(array $metadata): array
     {
         $return = [];
 
@@ -158,17 +171,13 @@ class DbSqliteFunctions extends DbBaseFunctions
         return $return;
     }
 
+    #[Override]
     public function getIsolationLevelCommand(?IsolationLevelEnum $isolationLevel = null): string
     {
-        switch ($isolationLevel) {
-            case IsolationLevelEnum::READ_UNCOMMITTED:
-                return "PRAGMA read_uncommitted = true;";
-            case IsolationLevelEnum::READ_COMMITTED:
-            case IsolationLevelEnum::REPEATABLE_READ:
-            case IsolationLevelEnum::SERIALIZABLE:
-            default:
-                return "";
-        }
+        return match ($isolationLevel) {
+            IsolationLevelEnum::READ_UNCOMMITTED => "PRAGMA read_uncommitted = true;",
+            default => "",
+        };
     }
     
 }
