@@ -2,13 +2,13 @@
 
 namespace ByJG\AnyDataset\Db;
 
-use ByJG\AnyDataset\Core\GenericIterator;
 use ByJG\AnyDataset\Db\Traits\PreFetchTrait;
+use ByJG\Serializer\PropertyHandler\PropertyHandlerInterface;
+use Override;
 use PDO;
 use PDOStatement;
-use ReturnTypeWillChange;
 
-class DbIterator extends GenericIterator
+class DbIterator extends GenericDbIterator
 {
     use PreFetchTrait;
 
@@ -18,40 +18,49 @@ class DbIterator extends GenericIterator
     private ?PDOStatement $statement;
 
     /**
+     * @var string|null
+     */
+    private ?string $entityClass;
+
+    /**
+     * @var PropertyHandlerInterface|null
+     */
+    private ?PropertyHandlerInterface $entityTransformer;
+
+    /**
      * @param PDOStatement $recordset
      * @param int $preFetch
+     * @param string|null $entityClass
+     * @param PropertyHandlerInterface|null $entityTransformer
      */
-    public function __construct(PDOStatement $recordset, int $preFetch = 0)
+    public function __construct(PDOStatement $recordset, int $preFetch = 0, ?string $entityClass = null, ?PropertyHandlerInterface $entityTransformer = null)
     {
         $this->statement = $recordset;
+        $this->entityClass = $entityClass;
+        $this->entityTransformer = $entityTransformer;
         $this->initPreFetch($preFetch);
     }
 
-    /**
-     * @return int
-     */
-    #[ReturnTypeWillChange]
-    public function count(): int
-    {
-        return $this->statement->rowCount();
-    }
-
+    #[Override]
     public function isCursorOpen(): bool
     {
         return !is_null($this->statement);
     }
 
+    #[Override]
     public function releaseCursor(): void
     {
         if ($this->isCursorOpen()) {
+            /** @psalm-suppress PossiblyNullReference */
             $this->statement->closeCursor();
             $this->statement = null;
         }
     }
 
+    #[Override]
     protected function fetchRow(): array|bool
     {
-        return $this->statement->fetch(PDO::FETCH_ASSOC);
+        return $this->statement?->fetch(PDO::FETCH_ASSOC) ?? false;
     }
 
     public function __destruct()
