@@ -4,11 +4,13 @@ namespace ByJG\AnyDataset\Db;
 
 use ByJG\AnyDataset\Core\Enum\Relation;
 use ByJG\AnyDataset\Core\IteratorFilterFormatter;
-use ByJG\AnyDataset\Db\Helpers\SqlHelper;
+use ByJG\AnyDataset\Db\SqlDialect\SqlHelper;
+use Override;
 
 class IteratorFilterSqlFormatter extends IteratorFilterFormatter
 {
-    public function format(array $filters, string $tableName = null, array &$params = [], string $returnFields = "*"): string
+    #[Override]
+    public function format(array $filters, ?string $tableName = null, array &$params = [], string $returnFields = "*"): string
     {
         $params = array();
 
@@ -18,7 +20,7 @@ class IteratorFilterSqlFormatter extends IteratorFilterFormatter
             $sql .= " where @@sqlFilter ";
         }
 
-        return SqlHelper::createSafeSQL(
+        return self::createSafeSQL(
             $sql,
             [
                 "@@returnFields" => $returnFields,
@@ -28,6 +30,7 @@ class IteratorFilterSqlFormatter extends IteratorFilterFormatter
         );
     }
 
+    #[Override]
     public function getRelation(string $name, Relation $relation, mixed $value, array &$param): string
     {
         $paramName = $name;
@@ -95,5 +98,11 @@ class IteratorFilterSqlFormatter extends IteratorFilterFormatter
             },
         };
 
-        return $data($param, $name, $paramName, $value);    }
+        return $data($param, $name, $paramName, $value);
+    }
+
+    public static function createSafeSQL(string $sql, array $list): string
+    {
+        return str_replace(array_keys($list), array_values($list), $sql);
+    }
 }
