@@ -3,14 +3,46 @@
 namespace ByJG\AnyDataset\Db;
 
 use ByJG\AnyDataset\Db\Exception\DbDriverNotConnected;
+use ByJG\AnyDataset\Db\SqlDialect\DblibDialect;
+use ByJG\AnyDataset\Db\SqlDialect\GenericPdoDialect;
+use ByJG\AnyDataset\Db\SqlDialect\MysqlDialect;
+use ByJG\AnyDataset\Db\SqlDialect\OciDialect;
+use ByJG\AnyDataset\Db\SqlDialect\PgsqlDialect;
+use ByJG\AnyDataset\Db\SqlDialect\SqliteDialect;
+use ByJG\AnyDataset\Db\SqlDialect\SqlsrvDialect;
 use ByJG\Util\Uri;
+use Override;
+use PDO;
 
 class PdoLiteral extends DbPdoDriver
 {
 
+    #[Override]
     public static function schema()
     {
         return null;
+    }
+
+    #[Override]
+    public function getSqlDialectClass(): string
+    {
+        // Detect PDO driver at runtime
+        $pdo = $this->getDbConnection();
+        if ($pdo instanceof PDO) {
+            $driver = $pdo->getAttribute(PDO::ATTR_DRIVER_NAME);
+
+            return match ($driver) {
+                'mysql' => MysqlDialect::class,
+                'sqlite' => SqliteDialect::class,
+                'pgsql' => PgsqlDialect::class,
+                'sqlsrv' => SqlsrvDialect::class,
+                'dblib' => DblibDialect::class,
+                'oci' => OciDialect::class,
+                default => GenericPdoDialect::class,
+            };
+        }
+
+        return GenericPdoDialect::class;
     }
 
     /**
@@ -33,6 +65,6 @@ class PdoLiteral extends DbPdoDriver
             $credential = "$username:$password@";
         }
 
-        parent::__construct(new Uri("literal://{$credential}{$parts[0]}?connection=" . urlencode($parts[1])), $preOptions, $postOptions, $executeAfterConnect);
+        parent::__construct(new Uri("literal://{$credential}{$parts[0]}?connection=" . urlencode($parts[1] ?? '')), $preOptions, $postOptions, $executeAfterConnect);
     }
 }
