@@ -3,14 +3,23 @@
 namespace ByJG\AnyDataset\Db;
 
 use ByJG\AnyDataset\Core\Exception\NotAvailableException;
+use ByJG\AnyDataset\Db\SqlDialect\SqlsrvDialect;
 use ByJG\Util\Uri;
+use Override;
 
 class PdoSqlsrv extends PdoDblib
 {
 
+    #[Override]
     public static function schema(): array
     {
         return ['sqlsrv'];
+    }
+
+    #[Override]
+    public function getSqlDialectClass(): string
+    {
+        return SqlsrvDialect::class;
     }
 
     /**
@@ -24,15 +33,16 @@ class PdoSqlsrv extends PdoDblib
         parent::__construct($connUri);
     }
 
+    #[Override]
     protected function getMssqlUri(Uri $connUri): Uri
     {
         /** @var Uri $uri */
-        $uri = Uri::getInstanceFromString("pdo://");
+        $uri = Uri::getInstance("pdo://");
 
         return $uri
-            ->withUserInfo($connUri->getUsername(), $connUri->getPassword())
+            ->withUserInfo($connUri->getUsername() ?? '', $connUri->getPassword())
             ->withHost($connUri->getScheme())
-            ->withQueryKeyValue("Server" , $connUri->getHost() . (!empty($connUri->getPort()) ? "," . $connUri->getPort() : ""))
+            ->withQueryKeyValue("Server", $connUri->getHost() . (!empty($connUri->getPort()) ? "," . (string)$connUri->getPort() : ""))
             ->withQueryKeyValue("Database", ltrim($connUri->getPath(), "/"))
             ->withQueryKeyValue('TrustServerCertificate', 'true')
         ;
