@@ -1,12 +1,13 @@
 <?php
 
-namespace ByJG\AnyDataset\Db\Helpers;
+namespace ByJG\AnyDataset\Db\SqlDialect;
 
 use ByJG\AnyDataset\Core\Exception\NotAvailableException;
-use ByJG\AnyDataset\Db\DbDriverInterface;
+use ByJG\AnyDataset\Db\DatabaseExecutor;
 use ByJG\AnyDataset\Db\IsolationLevelEnum;
+use Override;
 
-class DbDblibFunctions extends DbBaseFunctions
+class DblibDialect extends BaseSqlDialect
 {
 
     public function __construct()
@@ -17,6 +18,7 @@ class DbDblibFunctions extends DbBaseFunctions
         $this->deliTableRight = '"';
     }
 
+    #[Override]
     public function concat(string $str1, ?string $str2 = null): string
     {
         return implode(' + ', func_get_args());
@@ -30,6 +32,7 @@ class DbDblibFunctions extends DbBaseFunctions
      * @return string
      * @throws NotAvailableException
      */
+    #[Override]
     public function limit(string $sql, int $start, int $qty = 50): string
     {
         throw new NotAvailableException("DBLib does not support LIMIT feature.");
@@ -41,23 +44,27 @@ class DbDblibFunctions extends DbBaseFunctions
      * @param int $qty
      * @return string
      */
+    #[Override]
     public function top(string $sql, int $qty): string
     {
         if (stripos($sql, ' TOP ') === false) {
-            return  preg_replace("/^\\s*(select) /i", "\\1 top $qty ", $sql);
+            $result = preg_replace("/^\\s*(select) /i", "\\1 top $qty ", $sql);
+            return $result !== null ? $result : $sql;
         }
 
-        return preg_replace(
+        $result = preg_replace(
             '~(\s[Tt][Oo][Pp])\s.*?\d+\s~',
             '$1 ' . $qty . ' ',
             $sql
         );
+        return $result !== null ? $result : $sql;
     }
 
     /**
      * Return if the database provider have a top or similar function
      * @return bool
      */
+    #[Override]
     public function hasTop(): bool
     {
         return true;
@@ -67,6 +74,7 @@ class DbDblibFunctions extends DbBaseFunctions
      * Return if the database provider have a limit function
      * @return bool
      */
+    #[Override]
     public function hasLimit(): bool
     {
         return false;
@@ -80,6 +88,7 @@ class DbDblibFunctions extends DbBaseFunctions
      * @return string
      * @example $db->getDbFunctions()->SQLDate("d/m/Y H:i", "dtcriacao")
      */
+    #[Override]
     public function sqlDate(string $format, ?string $column = null): string
     {
         if (is_null($column)) {
@@ -112,6 +121,7 @@ class DbDblibFunctions extends DbBaseFunctions
         );
     }
 
+    #[Override]
     public function getSqlLastInsertId(): string
     {
         return "select @@identity id";
@@ -122,26 +132,29 @@ class DbDblibFunctions extends DbBaseFunctions
      * @return string
      * @throws NotAvailableException
      */
+    #[Override]
     public function forUpdate(string $sql): string
     {
         throw new NotAvailableException('FOR UPDATE not available for Mssql/Dblib');
     }
 
+    #[Override]
     public function hasForUpdate(): bool
     {
         return false;
     }
 
-    public function getTableMetadata(DbDriverInterface $dbdataset, string $tableName): array
+    #[Override]
+    public function getTableMetadata(DatabaseExecutor $executor, string $tableName): array
     {
         $sql = "select * from INFORMATION_SCHEMA.COLUMNS where TABLE_NAME = '$tableName'";
-        return $this->getTableMetadataFromSql($dbdataset, $sql);
+        return $this->getTableMetadataFromSql($executor, $sql);
     }
 
-    protected function parseColumnMetadata($metadata)
+    #[Override]
+    protected function parseColumnMetadata(array $metadata): array
     {
         $return = [];
-
 
         foreach ($metadata as $key => $value) {
             if (!empty($value['character_maximum_length'])) {
@@ -161,6 +174,7 @@ class DbDblibFunctions extends DbBaseFunctions
         return $return;
     }
 
+    #[Override]
     public function getIsolationLevelCommand(?IsolationLevelEnum $isolationLevel = null): string
     {
         return match ($isolationLevel) {
