@@ -2,17 +2,12 @@
 
 namespace ByJG\AnyDataset\Db;
 
-use ByJG\AnyDataset\Core\GenericIterator;
-use ByJG\AnyDataset\Lists\ArrayDataset;
-use DateInterval;
+use ByJG\Serializer\PropertyHandler\PropertyHandlerInterface;
 use Psr\SimpleCache\CacheInterface;
-use Psr\SimpleCache\InvalidArgumentException;
 
 class SqlStatement
 {
     protected string $sql;
-
-    protected array $cachedStatement = [];
 
     protected ?CacheInterface $cache = null;
 
@@ -20,27 +15,70 @@ class SqlStatement
 
     protected ?string $cacheKey = null;
 
-    public function __construct(string $sql)
+    protected array $params = [];
+
+    protected ?string $entityClass = null;
+
+    protected ?PropertyHandlerInterface $entityTransformer = null;
+
+    public function __construct(string $sql, ?array $params = [])
     {
         $this->sql = $sql;
         $this->cache = null;
         $this->cacheTime = null;
         $this->cacheKey = null;
+        $this->params = $params ?? [];
     }
 
-    public function withCache(CacheInterface $cache, string $cacheKey, int $cacheTime = 60): SqlStatement
+    public static function from(string $sql, ?array $params = []): static
     {
-        $this->cache = $cache;
-        $this->cacheTime = $cacheTime;
-        $this->cacheKey = $cacheKey;
-        return $this;
+        return new static($sql, $params);
     }
 
-    public function withoutCache()
+    public function withCache(CacheInterface $cache, string $cacheKey, int $cacheTime = 60): static
     {
-        $this->cache = null;
-        $this->cacheTime = null;
-        $this->cacheKey = null;
+        $statement = clone $this;
+        $statement->cache = $cache;
+        $statement->cacheTime = $cacheTime;
+        $statement->cacheKey = $cacheKey;
+        return $statement;
+    }
+
+    public function withoutCache(): static
+    {
+        $statement = clone $this;
+        $statement->cache = null;
+        $statement->cacheTime = null;
+        $statement->cacheKey = null;
+        return $statement;
+    }
+
+    public function withParams(?array $params): static
+    {
+        $statement = clone $this;
+        $statement->params = array_merge($this->params, $params ?? []);
+        return $statement;
+    }
+
+    public function withEntityClass(string $entityClass): static
+    {
+        $statement = clone $this;
+        $statement->entityClass = $entityClass;
+        return $statement;
+    }
+
+    public function withEntityTransformer(PropertyHandlerInterface $entityTransformer): static
+    {
+        $statement = clone $this;
+        $statement->entityTransformer = $entityTransformer;
+        return $statement;
+    }
+
+    public function withoutParams(): static
+    {
+        $statement = clone $this;
+        $statement->params = [];
+        return $statement;
     }
 
     public function getSql(): string
@@ -63,97 +101,18 @@ class SqlStatement
         return $this->cacheKey;
     }
 
-
-    public function getIterator(DbDriverInterface $dbDriver, ?array $param = [], int $preFetch = 0): GenericIterator
+    public function getParams(): ?array
     {
-        $cacheKey = "";
-        if (!empty($this->cache)) {
-            ksort($param);
-            $cacheKey = $this->cacheKey . ':' . md5(json_encode($param));
-            if ($this->cache->has($cacheKey)) {
-                return (new ArrayDataset($this->cache->get($cacheKey)))->getIterator();
-            }
-        }
-
-        do {
-            $lock = $this->mutexIsLocked($cacheKey);
-            if ($lock !== false) {
-                usleep(200);
-                continue;
-            }
-
-            $this->mutexLock($cacheKey);
-            try {
-                $statement = $dbDriver->prepareStatement($this->sql, $param, $this->cachedStatement);
-
-                $dbDriver->executeCursor($statement);
-                $iterator = $dbDriver->getIterator($statement, preFetch: $preFetch);
-
-                if (!empty($this->cache)) {
-                    $cachedItem = $iterator->toArray();
-                    $this->cache->set($cacheKey, $cachedItem, $this->cacheTime);
-                    return (new ArrayDataset($cachedItem))->getIterator();
-                }
-
-                return $iterator;
-            } finally {
-                $this->mutexRelease($cacheKey);
-            }
-        } while (true);
+        return $this->params;
     }
 
-    public function getScalar(DbDriverInterface $dbDriver, ?array $array = null): mixed
+    public function getEntityClass(): ?string
     {
-        $stmt = $dbDriver->prepareStatement($this->sql, $array, $this->cachedStatement);
-        $dbDriver->executeCursor($stmt);
-
-        return $dbDriver->getScalar($stmt);
+        return $this->entityClass;
     }
 
-    public function execute(DbDriverInterface $dbDriver, ?array $array = null): void
+    public function getEntityTransformer(): ?PropertyHandlerInterface
     {
-        $stmt = $dbDriver->prepareStatement($this->sql, $array, $this->cachedStatement);
-        $dbDriver->executeCursor($stmt);
-        $dbDriver->processMultiRowset($stmt);
+        return $this->entityTransformer;
     }
-
-    public function executeAndGetId(DbDriverInterface $dbDriver, ?array $array = null): mixed
-    {
-        return $dbDriver->getDbHelper()->executeAndGetInsertedId($dbDriver, $this->sql, $array);
-    }
-
-    /**
-     * @return mixed
-     * @throws InvalidArgumentException
-     */
-    protected function mutexIsLocked(string $cacheKey): mixed
-    {
-        if (empty($this->cache)) {
-            return false;
-        }
-        return $this->cache->get($cacheKey . ".lock", false);
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    protected function mutexLock(string $cacheKey): void
-    {
-        if (empty($this->cache)) {
-            return;
-        }
-        $this->cache->set($cacheKey . ".lock", time(), DateInterval::createFromDateString('5 min'));
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    protected function mutexRelease(string $cacheKey): void
-    {
-        if (empty($this->cache)) {
-            return;
-        }
-        $this->cache->delete($cacheKey . ".lock");
-    }
-
 }
