@@ -2,12 +2,22 @@
 
 namespace Test\Helpers;
 
-use ByJG\AnyDataset\Db\Helpers\DbSqlsrvFunctions;
+use ByJG\AnyDataset\Db\SqlDialect\SqlsrvDialect;
+use Override;
 
+/**
+ * Class DbSqlsrvFunctionsTest
+ *
+ * Tests for the SQLSRV database functions
+ */
 class DbSqlsrvFunctionsTest extends DbDblibFunctionsTest
 {
+    /**
+     * Set up the test object
+     */
+    #[Override]
     protected function setUp(): void
     {
-        $this->object = new DbSqlsrvFunctions();
+        $this->object = new SqlsrvDialect();
     }
 }
