@@ -17,6 +17,11 @@ class PdoOciTest extends BasePdo
 
     protected function createInstance()
     {
+        if (!extension_loaded('pdo_oci')) {
+            $this->testSkipped = true;
+            $this->markTestSkipped("PDO ORACLE OCI extension is not loaded");
+        }
+
         $this->escapeQuote = "''";
 
         $host = getenv('ORACLE_TEST_HOST');
@@ -41,7 +46,7 @@ class PdoOciTest extends BasePdo
     protected function createDatabase()
     {
         //create the database
-        $this->dbDriver->execute("CREATE TABLE Dogs (
+        $this->executor->execute("CREATE TABLE Dogs (
             Id INT GENERATED ALWAYS as IDENTITY(START with 1 INCREMENT by 1), 
             Breed varchar2(50), 
             Name varchar2(50), 
@@ -52,20 +57,20 @@ class PdoOciTest extends BasePdo
 
     public function deleteDatabase()
     {
-        $this->dbDriver->execute('drop table Dogs');
+        $this->executor->execute('drop table Dogs');
     }
 
     public function testGetDate() {
-        $data = $this->dbDriver->getScalar("SELECT TO_DATE('2018-07-26', 'YYYY-MM-DD') FROM DUAL ");
+        $data = $this->executor->getScalar("SELECT TO_DATE('2018-07-26', 'YYYY-MM-DD') FROM DUAL ");
         $this->assertEquals("26-JUL-18", $data);
 
-        $data = $this->dbDriver->getScalar("SELECT TO_TIMESTAMP('2018-07-26 20:02:03', 'YYYY-MM-DD HH24:MI:SS') FROM DUAL ");
+        $data = $this->executor->getScalar("SELECT TO_TIMESTAMP('2018-07-26 20:02:03', 'YYYY-MM-DD HH24:MI:SS') FROM DUAL ");
         $this->assertEquals("26-JUL-18 08.02.03.000000000 PM", $data);
     }
 
     public function testGetMetadata()
     {
-        $metadata = $this->dbDriver->getDbHelper()->getTableMetadata($this->dbDriver, 'Dogs');
+        $metadata = $this->executor->getHelper()->getTableMetadata($this->executor, 'Dogs');
 
         foreach ($metadata as $key => $field) {
             unset($metadata[$key]['dbType']);
