@@ -2,7 +2,9 @@
 
 namespace TestDb;
 
+use ByJG\AnyDataset\Db\DatabaseExecutor;
 use ByJG\AnyDataset\Db\Factory;
+use PDOException;
 
 class PdoPostgresTest extends BasePdo
 {
@@ -22,9 +24,10 @@ class PdoPostgresTest extends BasePdo
         }
 
         $dbDriver = Factory::getDbInstance("pgsql://postgres:$password@$host");
-        $exists = $dbDriver->getScalar('select count(1) from pg_catalog.pg_database where datname = \'test\'');
+        $executor = DatabaseExecutor::using($dbDriver);
+        $exists = $executor->getScalar('select count(1) from pg_catalog.pg_database where datname = \'test\'');
         if ($exists == 0) {
-            $dbDriver->execute('CREATE DATABASE test');
+            $executor->execute('CREATE DATABASE test');
         }
         return Factory::getDbInstance("pgsql://postgres:$password@$host/test");
     }
@@ -32,31 +35,31 @@ class PdoPostgresTest extends BasePdo
     protected function createDatabase()
     {
         //create the database
-        $this->dbDriver->execute("CREATE TABLE Dogs (Id SERIAL PRIMARY KEY, Breed VARCHAR(50), Name VARCHAR(50), Age INTEGER, Weight NUMERIC(10,2))");
+        $this->executor->execute("CREATE TABLE Dogs (Id SERIAL PRIMARY KEY, Breed VARCHAR(50), Name VARCHAR(50), Age INTEGER, Weight NUMERIC(10,2))");
     }
 
     public function deleteDatabase()
     {
-        $this->dbDriver->execute('drop table Dogs;');
+        $this->executor->execute('drop table Dogs;');
     }
 
     public function testGetDate() {
-        $data = $this->dbDriver->getScalar("SELECT CAST('2018-07-26' AS DATE) ");
+        $data = $this->executor->getScalar("SELECT CAST('2018-07-26' AS DATE) ");
         $this->assertEquals("2018-07-26", $data);
 
-        $data = $this->dbDriver->getScalar("SELECT CAST('2018-07-26 20:02:03' AS TIMESTAMP) ");
+        $data = $this->executor->getScalar("SELECT CAST('2018-07-26 20:02:03' AS TIMESTAMP) ");
         $this->assertEquals("2018-07-26 20:02:03", $data);
     }
 
     public function testDontParseParam_3() {
-        $this->expectException(\PDOException::class);
+        $this->expectException(PDOException::class);
         
         parent::testDontParseParam_3();
     }
 
     public function testGetMetadata()
     {
-        $metadata = $this->dbDriver->getDbHelper()->getTableMetadata($this->dbDriver, 'Dogs');
+        $metadata = $this->executor->getHelper()->getTableMetadata($this->executor, 'Dogs');
 
         foreach ($metadata as $key => $field) {
             unset($metadata[$key]['dbType']);
