@@ -17,7 +17,7 @@ class PdoSqliteTest extends BasePdo
             $this->host = "/tmp/test.db";
         }
 
-        $uri = Uri::getInstanceFromString("sqlite://" . $this->host);
+        $uri = Uri::getInstance("sqlite://" . $this->host);
 
         return Factory::getDbInstance($uri);
     }
@@ -25,7 +25,7 @@ class PdoSqliteTest extends BasePdo
     protected function createDatabase()
     {
         //create the database
-        $this->dbDriver->execute("CREATE TABLE Dogs (Id INTEGER NOT NULL PRIMARY KEY, Breed VARCHAR(50), Name VARCHAR(50), Age INTEGER, Weight NUMERIC(10,2))");
+        $this->executor->execute("CREATE TABLE Dogs (Id INTEGER NOT NULL PRIMARY KEY, Breed VARCHAR(50), Name VARCHAR(50), Age INTEGER, Weight NUMERIC(10,2))");
     }
 
     public function deleteDatabase()
@@ -35,14 +35,14 @@ class PdoSqliteTest extends BasePdo
 
     public function testGetAllFields()
     {
-        $this->markTestSkipped('Skipped: SqlLite does not support get all fields');
+        $this->markTestSkipped('SQLite does not support getAllFields() method');
     }
 
     public function testGetDate() {
-        $data = $this->dbDriver->getScalar("SELECT DATE('2018-07-26') ");
+        $data = $this->executor->getScalar("SELECT DATE('2018-07-26') ");
         $this->assertEquals("2018-07-26", $data);
 
-        $data = $this->dbDriver->getScalar("SELECT DATETIME('2018-07-26 20:02:03') ");
+        $data = $this->executor->getScalar("SELECT DATETIME('2018-07-26 20:02:03') ");
         $this->assertEquals("2018-07-26 20:02:03", $data);
     }
 }
